:art: update bootstrap.sh

+++ bootstrap.sh
@@ -10,6 +10,20 @@ cmd_missing() {
     !(type "$1" > /dev/null 2>&1)
 }
 
+realpath () {
+    f=$@;
+    if [ -d "$f" ]; then
+        base="";
+        dir="$f";
+    else
+        base="/$(basename "$f")";
+        dir=$(dirname "$f");
+    fi;
+    dir=$(cd "$dir" && /bin/pwd);
+    echo "$dir$base"
+}
+
+
 ###
 ##  1. install homebrew
 ###
@@ -29,7 +43,7 @@ declare -a ESSENTIALS=(
     "n"
     "rbenv"
     "goenv"
-    "zplug"
+    "nvim"
 )
 for cmd in "${ESSENTIALS[@]}"; do
     if cmd_missing $cmd; then brew install $cmd; fi
@@ -38,10 +52,28 @@ done
 ###
 ##  3. setup zplug
 ###
-zplug install
-
-$ZPLUG_HOME=/usr/local/opt/zplug
+ZPLUG_HOME=/usr/local/opt/zplug
+if [ ! -d $ZPLUG_HOME]; then
+    brew install zplug
+fi
 
 mkdir -p $HOME/.zfunctions
-ln -s $ZPLUG_HOME/repos/sindresorhus/pure/async.zsh $HOME/.zfunctions/async
-ln -s $ZPLUG_HOME/repos/sindresorhus/pure/pure.zsh $HOME/.zfunctions/prompt_pure_setup
+if [ ! -f $HOME/.zfunctions/async ]; then
+    ln -s $ZPLUG_HOME/repos/sindresorhus/pure/async.zsh $HOME/.zfunctions/async
+fi
+if [ ! -f $HOME/.zfunctions/prompt_pure_setup ]; then
+    ln -s $ZPLUG_HOME/repos/sindresorhus/pure/pure.zsh $HOME/.zfunctions/prompt_pure_setup
+fi
+
+###
+##  4. create symlinks
+###
+DOTFILES_DIR=$(dirname $0)
+for FILE in $(ls $DOTFILES_DIR/_*); do
+    FULLPATH=$(realpath $FILE)
+    DOTFILE=$(echo $FILE | sed -E 's/.*\/_(.+)$/.\1/')
+    if [ ! -f $HOME/$DOTFILE ]; then
+        ln -s $FULLPATH $HOME/$DOTFILE
+    fi
+done
+
